requirements.txt
python-telegram-bot>=22.0,<23.0
anthropic>=0.40
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.0
alembic>=1.13
